fix(journal): warn before closing the editor with unsaved work (#538)
The journal studio tracked `dirty` and drove its save button from it, but
registered no `beforeunload` handler. Closing the tab mid-entry threw the
work away without a word — while the identical action in the page builder
or the settings editor asked first.

## Why it was missing

Both of the other editors carried the guard as their own copy, byte for
byte identical. Nothing held them together, so the third editor simply
never got one, and no test would have said so.

Rather than add a third copy, the effect moves to `useUnsavedGuard(dirty)`
next to the existing `useScrollLock` — the same kind of small shared hook,
in the same place. A fourth editor now inherits the guard by using the
hook instead of by remembering to write it.

This is one of the four duplications #533 proposes to consolidate into a
shared save/dirty hook; taken out ahead of that refactor because it is
data loss and the fix is six lines.

## Verification

`tsc --noEmit` clean, `test:unit` 811 passed, `build` passes. The guard
itself cannot be asserted in a `node` test environment — `beforeunload`
needs a window — so this rests on the three call sites now sharing one
implementation that was already in production in two of them.

Closes #538
